requirements.txt
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0
rich>=13.7.0
pyperclip>=1.8.2
flask>=3.0.0
